Compliance with v7.1.0
