+++ resources/views/fields/tabs.blade.php
@@ -1,0 +1,1 @@
+<div>{{-- sponsor field --}}</div>
