+++ resources/views/form.blade.php
@@ -3,13 +3,11 @@
 @endif
 <x-tall-form :onKeyDownEnter="$onKeyDownEnter" :attr="[]">
     @include('tall-forms::includes.form-head')
-    <x-tall-attr :attr="\Tanthammar\TallForms\ConfigAttr::key('fields-wrapper')">
-    @foreach($fields as $field)
-        @if(filled($field))
-            @include('tall-forms::includes.field-root')
-        @endif
-    @endforeach
-    </x-tall-attr>
+    @if(isset($isTabs) && $isTabs)
+        @include('tall-forms::fields.tabs')
+    @else
+        @include('tall-forms::includes.field-loop')
+    @endif
     @include('tall-forms::includes.form-footer')
     @include('tall-forms::includes.buttons-root')
 </x-tall-form>
