+++ src/Commands/MakeForm.php
@@ -35,11 +35,11 @@ class MakeForm extends Command
             $use = "use Tanthammar\TallForms\Input;" . PHP_EOL;
             if(Str::contains($fields, 'Checkbox::make')) $use .= "use Tanthammar\TallForms\Checkbox;" . PHP_EOL;
             if(Str::contains($fields, 'KeyVal::make')) $use .= "use Tanthammar\TallForms\KeyVal;" . PHP_EOL;
-            if(Str::contains($fields, 'Group::make')) $use .= "use Tanthammar\TallForms\Group;" . PHP_EOL;
             if(Str::contains($fields, 'Repeater::make')) $use .= "use Tanthammar\TallForms\Repeater;" . PHP_EOL;
             if(Str::contains($fields, 'Textarea::make')) $use .= "use Tanthammar\TallForms\Textarea;" . PHP_EOL;
             if(Str::contains($fields, 'DatePicker::make')) $use .= "use Tanthammar\TallFormsSponsors\DatePicker;" . PHP_EOL;
             if(Str::contains($fields, 'Number::make')) $use .= "use Tanthammar\TallFormsSponsors\Number;" . PHP_EOL;
+            if(Str::contains($fields, 'Group::make')) $use .= "use Tanthammar\TallForms\Group;" . PHP_EOL;
             $stub = str_replace("use Tanthammar\TallForms\Input;", $use, $stub);
         }
