+++ src/Exceptions/InvalidArrayFieldType.php
@@ -11,6 +11,7 @@ class InvalidArrayFieldType extends \Exception
         $lookup = [
             'array' => 'Repeater',
             'keyval' => 'Keyval',
+            'tab' => 'Tab',
             'group' => 'Group'
         ];
