+++ src/ImageCropper.php
@@ -19,6 +19,7 @@ class ImageCropper extends BaseField
     public $align_label_top = true;
     public $allowed_in_repeater = false;
     public $allowed_in_keyval = false;
+    public $allowed_in_tab = false;
     public $allowed_in_group = false;
