+++ src/KeyVal.php
@@ -12,4 +12,6 @@ class KeyVal extends BaseField
     public $align_label_top = true;
     public $allowed_in_repeater = false;
     public $allowed_in_keyval = true;
+    public $allowed_in_tab = true;
+    public $allowed_in_group = true;
 }
