+++ src/Repeater.php
@@ -15,6 +15,8 @@ class Repeater extends BaseField
     public $align_label_top = true;
     public $allowed_in_repeater = false;
     public $allowed_in_keyval = false;
+    public $allowed_in_tab = false;
+    public $allowed_in_group = false;
 
     public function sortable(): self
     {
